Untested new adler32 with some ugly cast technique to make it return uint32_t on fastest possible way
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@3379 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ outputmessage.h
@@ -49,7 +49,7 @@ class OutputMessage : public NetworkMessage, boost::noncopyable
 		void addCryptoHeader(bool addChecksum)
 		{
 			if(addChecksum)
-				addHeader((uint32_t)(adlerChecksum((uint8_t*)(m_buffer + m_outputBufferStart), m_size)));
+				addHeader((adlerChecksum((uint8_t*)(m_buffer + m_outputBufferStart), m_size)));
 
 			addHeader((uint16_t)(m_size));
 		}
